Add a type module
