+++ src/message.ts
@@ -6,18 +6,7 @@ import {
   fuzzyRgx,
   continuationLineRgx,
 } from "./regex";
-
-export type Message = {
-  msgid: string;
-  msgidLine: number;
-  msgstr: string;
-  msgstrLine: number;
-  msgctxt: string;
-  msgctxtLine: number;
-  firstline: number;
-  lastline: number;
-  isfuzzy: boolean;
-};
+import { Message } from "./message_type";
 
 export function nextUntranslatedMessage(
   document: vscode.TextDocument,
