+++ src/message_type.ts
@@ -1,0 +1,11 @@
+export type Message = {
+  msgid: string;
+  msgidLine: number;
+  msgstr: string;
+  msgstrLine: number;
+  msgctxt: string;
+  msgctxtLine: number;
+  firstline: number;
+  lastline: number;
+  isfuzzy: boolean;
+};
